Now counts all extreme configurations including those with cycles.

+++ BrickCounting/RectilinearBrick.cpp
@@ -63,9 +63,9 @@ void RectilinearBrick::constructAllStronglyConnected(RectilinearBrick *bricks, i
     for(int8_t xx = x-3; xx <= x+3; ++xx) {
       bool xExtreme = xx == x-3 || xx == x+3;
       for(int8_t yy = y-1; yy <= y+1; ++yy) {
-	bool yExtreme = yy == y-1 || yy == y+1;
-	if(includeCorners || !(xExtreme && yExtreme))
-	  bricks[bricksSize++] = RectilinearBrick(xx, yy, level, true);
+        bool yExtreme = yy == y-1 || yy == y+1;
+        if(includeCorners || !(xExtreme && yExtreme))
+          bricks[bricksSize++] = RectilinearBrick(xx, yy, level, true);
       }
     }
   }
@@ -74,9 +74,9 @@ void RectilinearBrick::constructAllStronglyConnected(RectilinearBrick *bricks, i
     for(int8_t xx = x-1; xx <= x+1; ++xx) {
       bool xExtreme = xx == x-1 || xx == x+1;
       for(int8_t yy = y-3; yy <= y+3; ++yy) {
-	bool yExtreme = yy == y-3 || yy == y+3;
-	if(includeCorners || !(xExtreme && yExtreme))
-	  bricks[bricksSize++] = RectilinearBrick(xx, yy, level, false);
+        bool yExtreme = yy == y-3 || yy == y+3;
+        if(includeCorners || !(xExtreme && yExtreme))
+          bricks[bricksSize++] = RectilinearBrick(xx, yy, level, false);
       }
     }
   }
@@ -86,7 +86,7 @@ void RectilinearBrick::constructAllStronglyConnected(RectilinearBrick *bricks, i
     for(int8_t yy = y-2; yy <= y+2; ++yy) {
       bool yExtreme = yy == y-2 || yy == y+2;
       if(includeCorners || !(xExtreme && yExtreme))
-	bricks[bricksSize++] = RectilinearBrick(xx, yy, level, !horizontal());
+        bricks[bricksSize++] = RectilinearBrick(xx, yy, level, !horizontal());
     }
   }
 }
@@ -122,7 +122,7 @@ void RectilinearBrick::toLDR(std::ofstream &os, int x, int y, int ldrColor) cons
 
 std::ostream& operator<<(std::ostream& os, const RectilinearBrick& b)
 {
-  os<< (b.horizontal()?"[-":"[|") << " "<<((int)b.x)<<","<< ((int)b.y) <<", lv "<< b.level() << "]";
+  os<< (b.horizontal()?"[-":"[|") << " "<<((int)b.x)<<","<< ((int)b.y) <<", lv "<< (int)b.level() << "]";
   return os;
 }
